Project 4: Analysis of Marathon Results
In this project, We will be working with the dataset "marathon_results.csv". The dataset contains the age, sex, country of the participant. The dataset also contains the finish timings of the participants at different stages of the marathon, their finished positions overall and as per gender, and the division they belong.

Motivation: We want to study the effect of 1-D and 2-D Kernel Density Estimates on our prediction using Naive Bayes model. We want to see if the model's performance is better compared with K-NN model.

We want to study the above dataset and want to create a model in which we compute the 1-D Kernel Density Estimate of male and female runners using finish times, and use it together with the Bayes theorem to compute the probability that a runner with a given time was a female. We also want to implement the same model using the 2-D Kernel Density Estimate. 

We also want to compare accuracy of predictions obtained above to the predictions made using k-NN with the same input data.

Then we also want to use linear regression to predict finish times of runners based on their 5K times and explore any other interesting observation we make during the analysis of the dataset.